% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amazon_labeled}
\alias{amazon_labeled}
\title{Sentences from Amazon reviews}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 2 columns.}
\usage{
amazon_labeled
}
\description{
A dataset.
}
\keyword{datasets}
