% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean texts}
\usage{
clean_text(text)
}
\arguments{
\item{text}{Text source. A vector of character.}
}
\value{
A character vector.
}
\description{
Performs text cleaning.
}
