% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_corpus.R
\name{generate_corpus}
\alias{generate_corpus}
\title{GenerateS text corpus.}
\usage{
generate_corpus(text, type, sub_token, label = NULL, clean = TRUE)
}
\arguments{
\item{text}{A vector of character strings.}

\item{type}{Type of texts to be sampled. Possible values are texts, paragraphs, sentences, words, and characters.}

\item{sub_token}{A string specifying the text unit for filtering texts by length via \code{min_length} and \code{max_length}.}
}
\value{
Text corpus.
}
\description{
Generates and tokenizes a text corpus.
}
