% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imdb_labeled}
\alias{imdb_labeled}
\title{Sentences from IMDB reviews.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 748 rows and 2 columns.}
\usage{
imdb_labeled
}
\description{
A dataset.
}
\keyword{datasets}
