% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corpus.R
\name{load_amazon_sentences}
\alias{load_amazon_sentences}
\title{Loads amazon sentences}
\usage{
load_amazon_sentences(type, sub_token)
}
\arguments{
\item{type}{Tbd.}

\item{sub_token}{Tbd.}
}
\description{
Loads amazon sentences
}
