% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corpus.R
\name{load_imdb_sentences}
\alias{load_imdb_sentences}
\title{Loads IMDB sentences}
\usage{
load_imdb_sentences(type, sub_token)
}
\arguments{
\item{type}{Tbd.}

\item{sub_token}{Tbd.}
}
\description{
Loads IMDB sentences
}
