% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corpus.R
\name{load_twenty_newsgroups}
\alias{load_twenty_newsgroups}
\title{Loads 20 newsgroups}
\usage{
load_twenty_newsgroups(type, sub_token)
}
\arguments{
\item{type}{Tbd.}

\item{sub_token}{Tbd.}
}
\description{
Loads 20 newsgroups
}
