% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corpus.R
\name{load_yelp_sentences}
\alias{load_yelp_sentences}
\title{Loads yelp sentences}
\usage{
load_yelp_sentences(type, sub_token)
}
\arguments{
\item{type}{Tbd.}

\item{sub_token}{Tbd.}
}
\description{
Loads yelp sentences
}
