% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_text.R
\name{subset_text}
\alias{subset_text}
\title{Subset texts.}
\usage{
subset_text(corpus, min_length, max_length, word_list)
}
\arguments{
\item{corpus}{Text corpus.}

\item{min_length}{Minimum length of the texts to be sampled. \code{min_length} is an integer greater than 0. By default, \code{min_length} is set to 1.}

\item{max_length}{Maximum length of the texts to be sampled. \code{max_length} is an integer greater than 0. By default, \code{max_length} is set to 1.}

\item{word_list}{A word list.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Subset texts.
}
