% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twenty_newsgroups}
\alias{twenty_newsgroups}
\title{Twenty newsgroups dataset}
\format{An object of class \code{data.frame} with 11314 rows and 2 columns.}
\usage{
twenty_newsgroups
}
\description{
A dataset.
}
\keyword{datasets}
