% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corpus.R
\name{load_corpus}
\alias{load_corpus}
\title{Loads text corpus.}
\usage{
load_corpus(source, type, sub_token)
}
\arguments{
\item{source}{A string denoting the data source. Allowed values are.}

\item{type}{A string denoting the desired token type.}
}
\value{
Text corpus.
}
\description{
Decides upon a text corpus.
}
