% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corpus.R
\name{load_english_words}
\alias{load_english_words}
\title{Loads english words}
\usage{
load_english_words(type, sub_token)
}
\arguments{
\item{type}{Tbd.}

\item{sub_token}{Tbd.}
}
\description{
Loads english words
}
