% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corpus.R
\name{load_imdb_sentences}
\alias{load_imdb_sentences}
\title{Loads IMDB sentences}
\usage{
load_imdb_sentences(type, sub_token)
}
\arguments{
\item{type}{Type of texts to be sampled. Possible values are texts, paragraphs, sentences, words, and characters.}

\item{sub_token}{A string specifying the text unit for filtering texts by length via \code{min_length} and \code{max_length}.}
}
\value{
Text corpus.
}
\description{
Loads IMDB sentences
}
